# perf: improved hsv color mapping in HeatMapAnnotator by 20x