# Description

Replaced `np.zeros()` and subsequent fill operations with `np.full()`. This improves performance of the operation on a 1920x1080 frame by ~16x.

## Type of change
-   [x] Performance improvement

## How has this change been tested, please provide a testcase or example of how you tested the change?

Tested running in a loop using `time.perf_counter()` the default implementation with the updated one
Results are average over 10000 iterations
- with 3840x2160 frame: default takes average of 62ms, improved takes average of 5ms. Improvement of 11.5x
- with 1920x1080 frame: default takes average of 16ms, improved takes average 1ms. Improvement of 16x
- with 1280x720 frame: default takes average of 8ms, improved takes average 0.3ms. Improvement of 22.5x

All comparisons have been performed on the same hardware.

## Any specific deployment considerations

No additional requirements