@SkalskiP 2025-02-18
Hi @Armaggheddon 👋🏻 Thank you so much for your interest in `supervision`. It looks interesting. Could you share a script (preferably a Google Colab) where we could measure the performance boost?