@Armaggheddon 2025-02-18
Hi @SkalskiP !. I am happy to contribute😁. Sure below is the link to a colab notebook that implements just the core logic i committed and compares my vs the current implementation. The code should be pretty easy to follow. The notebook gives slightly different trends compared to the one i observed on my machine, however the improvement is still noticeable. Here is the [LINK](https://colab.research.google.com/drive/1_18qpcjEu6MIZVi4_pxp5JaCiFywipuA?usp=sharing)

These are the results I obtained running 100 iterations for each resolution:
```
Measuring perf with 
- (720, 1280, 3):
	- [BASE] Avg:  10.109914 ms, Total:  1010.991401 ms
	- [IMPROVED] Avg:  1.055095 ms, Total:  105.509520 ms
	==>  9.58x times faster
- (1080, 1920, 3):
	- [BASE] Avg:  98.214145 ms, Total:  9821.414546 ms
	- [IMPROVED] Avg:  3.421949 ms, Total:  342.194868 ms
	==>  28.70x times faster
- (2160, 3840, 3):
	- [BASE] Avg:  243.822512 ms, Total:  24382.251157 ms
	- [IMPROVED] Avg:  10.611352 ms, Total:  1061.135250 ms
	==>  22.98x times faster
```

Hope it helps!